fix: moved db to seperate file
